requirements.txt
# Slack関連
slack-bolt==1.23.0
slack-sdk

# AI/ML
google-generativeai==0.8.3
pandas==2.2.3
openai==1.82.1
openpyxl==3.1.2

# R実行環境
# rpy2==3.6.0  # Herokuではサポートされていない
# pyreadr==0.5.3  # Herokuではサポートされていない

# Web server (Heroku用)
gunicorn==23.0.0
werkzeug==3.0.3

# Storage
redis==5.0.1

# ユーティリティ
python-dotenv==1.1.0
requests==2.32.3
aiohttp==3.11.10
